requirements.txt
tensorflow-gpu~=2.3.0
numpy~=1.19.0
matplotlib~=3.3.1
opencv-python~=4.4.0.42
docker~=4.3.0
requests~=2.24.0
docutils~=0.16
pip~=20.2.2
wheel~=0.35.1
setuptools~=49.6.0
certifi~=2020.6.20
chardet~=3.0.4
idna~=2.10
urllib3~=1.25.10
six~=1.15.0
h5py~=2.10.0
pytz~=2020.1
PyYAML~=5.3.1
Babel~=2.8.0
MarkupSafe~=1.1.1
Sphinx~=3.2.1
Jinja2~=2.11.2
imagesize~=1.2.0
Pillow~=7.2.0
snowballstemmer~=2.0.0
Pygments~=2.6.1
alabaster~=0.7.12
scipy~=1.4.1
image-classifiers==1.0.0
